Observaciones Meteorológicas

Datos mensuales de temperatura y precipitación en estaciones meteorológicas en Argentina.

Uso

datos_nh_mensual

Formato

Un data.table con 381 filas y 10 columnas

mes

Fecha en formato %Y-%m-%d

precipitacion_mensual

Suma de la precipitación durante el mes en mm.

temperatura_media_mensual

Promedio de la temperatura en el mes en grados celsius.

codigo_nh

Código de identificación de la estación

estacion

Nombre de la estación.

provincia

Provincia donde se ubica la estación.

organismo

Organismo a cargo de la estación.

lat

Latitud.

lon

Longitud.

altura

Altura sobre el nivel del mar donde se ubica la estación.